pyproject.toml
[project]
name = "body-height-estimator"
version = "0.1.0"
description = "Add your description here"
authors = [{ name = "kittchy", email = "drums.radwimps203@gmail.com" }]
dependencies = [
    "wandb>=0.17.4",
    "sh>=2.0.7",
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "lightning>=2.0.0",
    "torchmetrics>=0.11.4",
    "hydra-core==1.3.2",
    "hydra-colorlog==1.2.0",
    "hydra-optuna-sweeper==1.2.0",
    "rootutils>=1.0.7",
    "pre-commit>=3.7.1",
    "rich>=13.7.1",
    "pytest>=8.2.2",
    "dataclass-argparse>=0.1.2",
    "pydantic>=2.8.2",
    "opencv-python>=4.10.0.84",
    "mediapipe>=0.10.14",
    "transformers>=4.42.4",
    "numpy<2.0",
    "pillow>=10.4.0",
    "rembg>=2.0.57",
]
readme = "README.md"
requires-python = ">= 3.10"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/"]

[tool.rye.scripts]

train = ["python", "src/train.py"]
eval = ["python", "src/train.py"]
preprocess = ["python", "src/preprocess.py"]
